Isolating JSON parsing exception in jsonParse
In previous code, the entire callback along with JSON parsing was
encapsulated with a try / catch block.  Thus any error that occurs
within the callback bubbles up to and is caught in this scope.

I added a unit test to specifically test for this in
`test\unit\utils\jsonParseTest.js` called `should call the callback only
once if an error is thrown in callback execution`.